♻️ ✨ *: major refactor & add tenant member entities

+++ libs/api/bll/src/features/actors/actor-images/actor-images.service.ts
@@ -3,9 +3,9 @@ import { HasuraService } from '../../../global/graphql/hasura.service';
 import { LogsService } from '../../../global/logs/logs.service';
 import { BadRequestException, ForbiddenException, Injectable, NotFoundException, Logger } from '@nestjs/common';
 
-import { Actor, ActorImageRepository } from '@okampus/api/dal';
-import { ActorImageType, EntityName } from '@okampus/shared/enums';
-import { mergeUnique, canAdminDelete, canAdminManage } from '@okampus/shared/utils';
+import { ActorImageRepository } from '@okampus/api/dal';
+import { EntityName } from '@okampus/shared/enums';
+import { mergeUnique } from '@okampus/shared/utils';
 
 import { EntityManager } from '@mikro-orm/core';
 
@@ -36,8 +36,7 @@ export class ActorImagesService extends RequestContext {
 
   async checkPermsCreate(props: ActorImageInsertInput) {
     if (Object.keys(props).length === 0) throw new BadRequestException('Create props cannot be empty.');
-    const requesterRoles = this.requester().adminRoles.getItems();
-    if (requesterRoles.some((adminRole) => canAdminManage(adminRole, { tenant: this.tenant() }))) return true;
+    
 
     // Custom logic
     return false;
@@ -45,8 +44,7 @@ export class ActorImagesService extends RequestContext {
 
   async checkPermsDelete(actorImage: ActorImage) {
     if (actorImage.deletedAt) throw new NotFoundException(`ActorImage was deleted on ${actorImage.deletedAt}.`);
-    const requesterRoles = this.requester().adminRoles.getItems();
-    if (requesterRoles.some((adminRole) => canAdminDelete(adminRole, actorImage))) return true;
+    
 
     // Custom logic
     return false;
@@ -56,9 +54,7 @@ export class ActorImagesService extends RequestContext {
     if (Object.keys(props).length === 0) throw new BadRequestException('Update props cannot be empty.');
 
     if (actorImage.deletedAt) throw new NotFoundException(`ActorImage was deleted on ${actorImage.deletedAt}.`);
-    if (actorImage.hiddenAt) throw new NotFoundException('ActorImage must be unhidden before it can be updated.');
-    const requesterRoles = this.requester().adminRoles.getItems();
-    if (requesterRoles.some((adminRole) => canAdminManage(adminRole, actorImage))) return true;
+    
 
     // Custom logic
     return actorImage.createdBy?.id === this.requester().id;
@@ -66,6 +62,7 @@ export class ActorImagesService extends RequestContext {
 
   async checkPropsConstraints(props: ActorImageSetInput) {
     this.hasuraService.checkForbiddenFields(props);
+    
 
     // Custom logic
     return true;
@@ -73,13 +70,20 @@ export class ActorImagesService extends RequestContext {
 
   async checkCreateRelationships(props: ActorImageInsertInput) {
     // Custom logic
-    props.tenantId = this.tenant().id;
+    
     props.createdById = this.requester().id;
+
+    
+    
 
     return true;
   }
 
-  async insertActorImageOne(selectionSet: string[], object: ActorImageInsertInput, onConflict?: ActorImageOnConflict) {
+  async insertActorImageOne(
+    selectionSet: string[],
+    object: ActorImageInsertInput,
+    onConflict?: ActorImageOnConflict,
+  ) {
     const canCreate = await this.checkPermsCreate(object);
     if (!canCreate) throw new ForbiddenException('You are not allowed to insert ActorImage.');
 
@@ -89,20 +93,12 @@ export class ActorImagesService extends RequestContext {
     const areRelationshipsValid = await this.checkCreateRelationships(object);
     if (!areRelationshipsValid) throw new BadRequestException('Relationships are not valid.');
 
-    selectionSet = mergeUnique(selectionSet, ['id', 'type', 'image.url', 'actor.id']);
+    selectionSet = mergeUnique(selectionSet, ['id']);
     const data = await this.hasuraService.insertOne('insertActorImageOne', selectionSet, object, onConflict);
 
     const actorImage = await this.actorImageRepository.findOneOrFail(data.insertActorImageOne.id);
     await this.logsService.createLog(EntityName.ActorImage, actorImage);
 
-    const type = data.insertActorImageOne.type;
-    if (type === ActorImageType.Avatar || type === ActorImageType.Banner) {
-      const actor = await this.em.findOne(Actor, { id: data.insertActorImageOne.actor.id });
-      if (actor) {
-        actor[type === ActorImageType.Avatar ? 'avatar' : 'banner'] = data.insertActorImageOne.image.url;
-        await this.em.persistAndFlush(actor);
-      }
-    }
     // Custom logic
     return data.insertActorImageOne;
   }
@@ -120,9 +116,12 @@ export class ActorImagesService extends RequestContext {
     return data.actorImage;
   }
 
-  async findActorImageByPk(selectionSet: string[], id: string) {
+  async findActorImageByPk(
+    selectionSet: string[],
+     id: string, 
+  ) {
     // Custom logic
-    const data = await this.hasuraService.findByPk('actorImageByPk', selectionSet, { id });
+    const data = await this.hasuraService.findByPk('actorImageByPk', selectionSet, {  id,  });
     return data.actorImageByPk;
   }
 
@@ -142,7 +141,7 @@ export class ActorImagesService extends RequestContext {
       if (!areRelationshipsValid) throw new BadRequestException('Create relationships are not valid.');
     }
 
-    selectionSet = [...selectionSet.filter((field) => field !== 'returning.id'), 'returning.id'];
+    selectionSet = mergeUnique(selectionSet, ['returning.id']);
     const data = await this.hasuraService.insert('insertActorImage', selectionSet, objects, onConflict);
 
     for (const inserted of data.insertActorImage.returning) {
@@ -154,41 +153,43 @@ export class ActorImagesService extends RequestContext {
     return data.insertActorImage;
   }
 
-  async updateActorImageMany(selectionSet: string[], updates: Array<ActorImageUpdates>) {
+  async updateActorImageMany(
+    selectionSet: string[],
+    updates: Array<ActorImageUpdates>,
+  ) {
     const areWheresCorrect = this.hasuraService.checkUpdates(updates);
     if (!areWheresCorrect) throw new BadRequestException('Where must only contain { id: { _eq: <id> } } in updates.');
 
     const actorImages = await this.actorImageRepository.findByIds(updates.map((update) => update.where.id._eq));
 
-    await Promise.all(
-      updates.map(async (update) => {
-        const actorImage = actorImages.find((actorImage) => actorImage.id === update.where.id._eq);
-        if (!actorImage) throw new NotFoundException(`ActorImage (${update.where.id._eq}) was not found.`);
+    await Promise.all(updates.map(async (update) => {
+      const actorImage = actorImages.find((actorImage) => actorImage.id === update.where.id._eq);
+      if (!actorImage) throw new NotFoundException(`ActorImage (${update.where.id._eq}) was not found.`);
 
-        const canUpdate = await this.checkPermsUpdate(update._set, actorImage);
-        if (!canUpdate)
-          throw new ForbiddenException(`You are not allowed to update ActorImage (${update.where.id._eq}).`);
+      const canUpdate = await this.checkPermsUpdate(update._set, actorImage);
+      if (!canUpdate) throw new ForbiddenException(`You are not allowed to update ActorImage (${update.where.id._eq}).`);
 
-        const arePropsValid = await this.checkPropsConstraints(update._set);
-        if (!arePropsValid) throw new BadRequestException(`Props are not valid in ${JSON.stringify(update._set)}.`);
-      }),
-    );
+      const arePropsValid = await this.checkPropsConstraints(update._set);
+      if (!arePropsValid) throw new BadRequestException(`Props are not valid in ${JSON.stringify(update._set)}.`);
+    }));
 
     const data = await this.hasuraService.updateMany('updateActorImageMany', selectionSet, updates);
 
-    await Promise.all(
-      actorImages.map(async (actorImage) => {
-        const update = updates.find((update) => update.where.id._eq === actorImage.id);
-        if (!update) return;
-        await this.logsService.updateLog(EntityName.ActorImage, actorImage, update._set);
-      }),
-    );
+    await Promise.all(actorImages.map(async (actorImage) => {
+      const update = updates.find((update) => update.where.id._eq === actorImage.id)
+      if (!update) return;
+      await this.logsService.updateLog(EntityName.ActorImage, actorImage, update._set);
+    }));
 
     // Custom logic
     return data.updateActorImageMany;
   }
 
-  async updateActorImageByPk(selectionSet: string[], pkColumns: ActorImagePkColumnsInput, _set: ActorImageSetInput) {
+  async updateActorImageByPk(
+    selectionSet: string[],
+    pkColumns: ActorImagePkColumnsInput,
+    _set: ActorImageSetInput,
+  ) {
     const actorImage = await this.actorImageRepository.findOneOrFail(pkColumns.id);
 
     const canUpdate = await this.checkPermsUpdate(_set, actorImage);
@@ -205,55 +206,42 @@ export class ActorImagesService extends RequestContext {
     return data.updateActorImageByPk;
   }
 
-  async deleteActorImage(selectionSet: string[], where: ActorImageBoolExp) {
-    const getData = await this.hasuraService.get('actorImage', ['id'], where);
-    const actorImagesData = getData.actorImage;
+  async deleteActorImage(
+    selectionSet: string[],
+    where: ActorImageBoolExp,
+  ) {
+    const isWhereCorrect = this.hasuraService.checkDeleteWhere(where);
+    if (!isWhereCorrect) throw new BadRequestException('Where must only contain { id: { _in: <Array<id>> } } in delete.');
 
-    if (!Array.isArray(actorImagesData) || actorImagesData.length === 0)
-      throw new BadRequestException('No corresponding ActorImage.');
+    const actorImages = await this.actorImageRepository.findByIds(where.id._in);
 
-    const actorImages = await this.actorImageRepository.findByIds(
-      actorImagesData.map(({ id }) => id),
-      { populate: ['actor'] },
-    );
+    await Promise.all(actorImages.map(async (actorImage) => {
+      const canDelete = await this.checkPermsDelete(actorImage);
+      if (!canDelete) throw new ForbiddenException(`You are not allowed to delete ActorImage (${actorImage.id}).`);
+    }));
 
-    await Promise.all(
-      actorImages.map(async (actorImage) => {
-        const canDelete = await this.checkPermsDelete(actorImage);
-        if (!canDelete) throw new ForbiddenException(`You are not allowed to delete ActorImage (${actorImage.id}).`);
-      }),
-    );
+    const data = await this.hasuraService.update('updateActorImage', selectionSet, where, { deletedAt: new Date().toISOString() });
 
-    const data = await this.hasuraService.update('updateActorImage', selectionSet, where, {
-      deletedAt: new Date().toISOString(),
-    });
-
-    await Promise.all(
-      actorImages.map(async (actorImage) => {
-        console.log(actorImage, actorImage.actor, actorImage.type);
-        if (actorImage.type === ActorImageType.Avatar || actorImage.type === ActorImageType.Banner)
-          actorImage.actor[actorImage.type === ActorImageType.Avatar ? 'avatar' : 'banner'] = '';
-        console.log(actorImage, actorImage.actor, actorImage.type);
-        await this.logsService.deleteLog(EntityName.ActorImage, actorImage.id);
-      }),
-    );
+    await Promise.all(actorImages.map(async (actorImage) => {
+      await this.logsService.deleteLog(EntityName.ActorImage, actorImage.id);
+    }));
 
     // Custom logic
     return data.updateActorImage;
   }
 
-  async deleteActorImageByPk(selectionSet: string[], id: string) {
+  async deleteActorImageByPk(
+    selectionSet: string[],
+    id: string,
+  ) {
     const actorImage = await this.actorImageRepository.findOneOrFail(id);
 
     const canDelete = await this.checkPermsDelete(actorImage);
     if (!canDelete) throw new ForbiddenException(`You are not allowed to delete ActorImage (${id}).`);
 
-    const data = await this.hasuraService.updateByPk(
-      'updateActorImageByPk',
-      selectionSet,
-      { id },
-      { deletedAt: new Date().toISOString() },
-    );
+    const data = await this.hasuraService.updateByPk('updateActorImageByPk', selectionSet, { id }, {
+      deletedAt: new Date().toISOString(),
+    });
 
     await this.logsService.deleteLog(EntityName.ActorImage, id);
     // Custom logic
@@ -266,18 +254,10 @@ export class ActorImagesService extends RequestContext {
     orderBy?: Array<ActorImageOrderBy>,
     distinctOn?: Array<ActorImageSelectColumn>,
     limit?: number,
-    offset?: number,
+    offset?: number
   ) {
     // Custom logic
-    const data = await this.hasuraService.aggregate(
-      'actorImageAggregate',
-      selectionSet,
-      where,
-      orderBy,
-      distinctOn,
-      limit,
-      offset,
-    );
+    const data = await this.hasuraService.aggregate('actorImageAggregate', selectionSet, where, orderBy, distinctOn, limit, offset);
     return data.actorImageAggregate;
   }
 }
